+++ examples/stream_weaver/motion_patterns.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+require 'slim_graph_r/stream_weaver'
+
+motion_pattern :fan_in_queue, style: :ruby
+motion_pattern :paired_policy_trace, style: :ruby
+motion_pattern :secure_paved_road, style: :ruby