+++ lib/slim_graph_r/stream_weaver_opal.rb
@@ -40,6 +40,12 @@ module SlimGraphR
       components << component
       component
     end
+
+    def motion_pattern(type, **options)
+      component = StreamWeaverComponent.new(SlimGraphR.motion_pattern(type, **options))
+      components << component
+      component
+    end
   end
 end